factor out treeFromRecords
