+++ uiCommon/ide/errors.ts
@@ -1,8 +1,8 @@
 import { AbstractInterpreter } from "../../core/abstractInterpreter";
 import { Rec, Int, int } from "../../core/types";
 import { Span, dlToSpan } from "./types";
-import { treeFromRecords } from "../visualizations/tree";
 import { mapTree, filterTree, getLeaves } from "../../util/tree";
+import { treeFromRecords } from "../generic/treeFromRecords";
 
 // just `TypeError` is a builtin
 export type DLTypeError = {
