+++ uiCommon/visualizations/tree.tsx
@@ -1,11 +1,10 @@
 import React, { useState } from "react";
-import { Tree } from "../../util/tree";
-import { Int, Rec, Res, StringLit, Term } from "../../core/types";
+import { Rec, StringLit, Term } from "../../core/types";
 import { VizTypeSpec } from "./typeSpec";
 import { AbstractInterpreter } from "../../core/abstractInterpreter";
 import { emptyCollapseState, TreeView } from "../generic/treeView";
 import { BareTerm } from "../dl/replViews";
-import { ppt } from "../../core/pretty";
+import { treeFromRecords } from "../generic/treeFromRecords";
 
 export const tree: VizTypeSpec = {
   name: "Tree",
@@ -47,38 +46,9 @@ function TreeViz(props: {
   }
 }
 
-type TermGraph = { [parentTerm: string]: Res[] };
-
 // currently dead
 // TODO: use this for "viz suggestions"
 export function canTreeViz(rec: Rec): boolean {
   const fields = Object.keys(rec.attrs);
   return fields.includes("id") && fields.includes("parent");
-}
-
-// query: something{id: C, parent: P}.
-// TODO: maybe specify a variable, and grab that binding?
-//   showing whole record is a bit noisy
-export function treeFromRecords(records: Res[], rootTerm: Term): Tree<Res> {
-  const graph: TermGraph = {};
-  records.forEach((res) => {
-    const parentID = (res.bindings.ParentID as Int).val;
-    const newChildren = [...(graph[parentID] || []), res];
-    graph[parentID] = newChildren;
-  });
-  return mkTree(graph, rootTerm, null);
-}
-
-function mkTree(
-  termGraph: TermGraph,
-  curID: Term,
-  curRes: Res | null
-): Tree<Res> {
-  return {
-    key: curID.toString(),
-    item: curRes,
-    children: (termGraph[ppt(curID)] || []).map((child) =>
-      mkTree(termGraph, child.bindings.ID, child)
-    ),
-  };
 }
